seperate makeQueue function
